Version 1.0
Version 1 compilador
Fase lexico

+++ EstructuraToken.h
@@ -16,7 +16,7 @@ struct Token{
     char nombre[80];
     enum TipoToken tipo;
     char lexema[80];
-    int valor;
+    double valor;
 };
 
 struct nodo
